requirements.txt
Django==4.2.20
mysqlclient>=2.1
django-environ==0.12
django-debug-toolbar==6.0
django-browser-reload==1.18.0
django-tailwind[reload]==4.2